refactor: moved useSessionData hook to /hooks directory
