+++ tsconfig.json
@@ -26,6 +26,7 @@
       "@/actions/*": ["./actions/*"],
       "@/api/*": ["./app/api/*"],
       "@/typings/*": ["./typings/*"],
+      "@/hooks/*": ["./hooks/*"],
     },
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
